requirements.txt
numpy==1.12.0
opencv-python==3.2.0.6
Pillow==4.0.0
pywin32==220
